fixed #2059: Fix complication errors on Android.

+++ scripting/javascript/bindings/ScriptingCore.cpp
@@ -1530,7 +1530,7 @@ jsval c_string_to_jsval(JSContext* cx, const char* v, size_t length /* = -1 */) 
         return JSVAL_NULL;
     }
     jsval ret = JSVAL_NULL;
-    size_t utf16_size = 0;
+    int utf16_size = 0;
     jschar* strUTF16 = (jschar*)cc_utf8_to_utf16(v, length, &utf16_size);
 
     if (strUTF16 && utf16_size > 0) {
